v0.2.0

+++ src/main.rs
@@ -433,7 +433,7 @@ fn print_thresholds(read_lengths: &Vec<u64>, read_qualities: &Vec<f32>, reads: &
     for k in qkeys.iter() {
         let data = &q_threshold_data[&k.to_string()];
 
-        let percent_reads = data[0] / reads;
+        let percent_reads = (data[0] / reads) as f64;
         let nreads = data[0].to_formatted_string(&Locale::en);
         let bp = data[1].to_formatted_string(&Locale::en);
 
@@ -447,7 +447,7 @@ fn print_thresholds(read_lengths: &Vec<u64>, read_qualities: &Vec<f32>, reads: &
     for k in lkeys.iter() {
         let data = &l_threshold_data[&k.to_string()];
 
-        let percent_reads: f64 = data[0] / reads;
+        let percent_reads: f64 = (data[0] / reads) as f64;
         let nreads = data[0].to_formatted_string(&Locale::en);
         let bp = data[1].to_formatted_string(&Locale::en);
